refactor modals with store updates & move to modal directory.
